pyproject.toml
[tool.poetry]
name = "metrics_tracker"
version = "0.1.0"
description = ""
authors = ["Max Shkurygin <max.atreides@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.3"
django-crispy-forms = "^1.10.0"
crispy_tailwind = "^0.2.0"
matplotlib = "^3.3.3"
gunicorn = "^20.0.4"
django-environ = "^0.4.5"
whitenoise = "^5.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
pylint = "^2.6.0"
rope = "^0.18.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
